webscraping using scrapy/requirement.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.26.9
botocore==1.29.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.19.6
proxyscrape==0.3.0
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==304
pywinpty==2.0.9
pyzmq==24.0.1
queuelib==1.6.2
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
scraperapi-sdk==0.2.2
Scrapy==2.7.1
scrapy-proxies==0.4
scrapy-proxy-pool==0.1.9
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
seaborn==0.12.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2022.6
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
w3lib==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.2
